Adding the remaining training and study code

+++ EventFeedbackAPI/Controllers/ParticipantController.cs
@@ -7,9 +7,9 @@ using System.Threading.Tasks;
 using EventFeedbackAPI.Application.dto;
 using EventFeedbackAPI.Infra.Ioc;
 using EventFeedbackAPI.Application.interfaces;
+using EventFeedbackAPI.Domain.authentication;
+using Microsoft.AspNetCore.Authorization;
  
-
-
 namespace EventFeedbackAPI.Controllers
 {
     [ApiController]
@@ -17,26 +17,46 @@ namespace EventFeedbackAPI.Controllers
     public class ParticipantController : Controller
     {
         private readonly IParticipantService _service;
+        private readonly IAuthenticate _authenticateService;
 
-        public ParticipantController(IParticipantService service)
+        public ParticipantController(IParticipantService service, IAuthenticate authenticateService)
         {
             _service = service;
+            _authenticateService = authenticateService;
         }
 
-
+            
         [HttpPost]
         public async Task<ActionResult> Create(ParticipantDto participant)
         {
+            if(participant == null)
+            {
+                return BadRequest("Invalid data!");
+            }
+
+            var participantExists = await _authenticateService.participantExists(participant.Cpf);
+
+            if (participantExists)
+            {
+                return BadRequest("This CPF is already registered.");
+            }
+
             var participantDtoIncluded = await _service.Create(participant);
             if (participantDtoIncluded == null)
             {
                 return BadRequest("An unexpected error occurred.");
             }
 
+            var token = _authenticateService.GenerateToken(participantDtoIncluded.Id, participantDtoIncluded.Cpf);
+           
+
             return Ok(participantDtoIncluded);
         }
 
+
+
         [HttpPut]
+        [Authorize]
         public async Task<ActionResult> Update(ParticipantDto participant)
         {
   
@@ -50,8 +70,17 @@ namespace EventFeedbackAPI.Controllers
         }
 
         [HttpDelete]
+        [Authorize]
         public async Task<ActionResult> Delete(int id)
         {
+            var participantId = int.Parse(User.FindFirst("id").Value);
+            var participant = await _service.FindAsync(participantId);
+
+            if(!participant.IsAdmin)
+            {
+                return Unauthorized("You are not allowed to perform this action.");
+            }
+
             var participantDtoDeleted = await _service.Delete(id);
             if (participantDtoDeleted == null)
             {
@@ -62,6 +91,7 @@ namespace EventFeedbackAPI.Controllers
         }
 
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult> Find(int id)
         {
             var participantDto = await _service.FindAsync(id);
